Fix reverse tunnel connect/data handler race condition
There was a problem whereby on a reverse tunnel, the server process
would trigger the connection to start on the client. However, if the
client was a bit slow to connect (as the attempt is a non-blocking
one), the server is unaware. It would then pass on data it read on the
server side, down the tunnel, for the client to process.

The client would try to perform that data handling on a TCP connection
which was not yet established, and errors would result.

Fix this by utilising the R2TCMD_RCONN message, in reverse, to signal
to the server that the client connected the reverse tunnel first. The
server would ignore read data for the tunnel until the client
confirmed connection.